refactor(mobile): extract widgets from channel_detail_page, dedup uploads
Structural polish pass:
- Extract message_actions.dart (action sheet, edit, delete, reaction row)
- Extract members_sheet.dart (member list, role management, avatars)
- Extract manage_channel_sheet.dart (join/leave, canvas, context cards)
- channel_detail_page.dart reduced from 1877 to 1047 lines (-44%)
- Deduplicate pickAndUploadAttachment/pickAndUploadVideo into shared
  pickAndUpload helper in compose_bar.dart
- Remove dead imports after extraction
