+++ mobile/lib/features/channels/channel_detail_page.dart
@@ -1,7 +1,6 @@
 import 'dart:async';
 
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:lucide_icons_flutter/lucide_icons.dart';
@@ -14,14 +13,16 @@ import '../profile/user_cache_provider.dart';
 import '../profile/user_profile.dart';
 import '../forum/forum_posts_view.dart';
 import 'channel.dart';
-import 'date_formatters.dart';
-import 'day_divider.dart';
-import 'emoji_picker.dart';
 import 'channel_management_provider.dart';
 import 'channel_messages_provider.dart';
 import 'channel_typing_provider.dart';
 import 'channels_provider.dart';
 import 'compose_bar.dart';
+import 'date_formatters.dart';
+import 'day_divider.dart';
+import 'manage_channel_sheet.dart';
+import 'members_sheet.dart';
+import 'message_actions.dart';
 import 'message_content.dart';
 import 'send_message_provider.dart';
 import 'thread_detail_page.dart';
@@ -118,7 +119,7 @@ class ChannelDetailPage extends HookConsumerWidget {
                 context: context,
                 isScrollControlled: true,
                 showDragHandle: true,
-                builder: (_) => _MembersSheet(
+                builder: (_) => MembersSheet(
                   channel: resolvedChannel,
                   currentPubkey: currentPubkey,
                 ),
@@ -134,7 +135,7 @@ class ChannelDetailPage extends HookConsumerWidget {
                   context: context,
                   isScrollControlled: true,
                   showDragHandle: true,
-                  builder: (_) => _ManageChannelSheet(channel: resolvedChannel),
+                  builder: (_) => ManageChannelSheet(channel: resolvedChannel),
                 );
                 if (shouldClose == true && context.mounted) {
                   Navigator.of(context).pop();
@@ -610,7 +611,7 @@ class _MessageBubble extends ConsumerWidget {
 
     return GestureDetector(
       behavior: HitTestBehavior.opaque,
-      onLongPress: () => _showMessageActions(
+      onLongPress: () => showMessageActions(
         context: context,
         ref: ref,
         message: message,
@@ -817,259 +818,6 @@ class _ReactionRow extends StatelessWidget {
 }
 
 // ---------------------------------------------------------------------------
-// Message actions (long-press sheet)
-// ---------------------------------------------------------------------------
-
-const _quickEmojis = ['👍', '❤️', '😂', '🎉', '👀', '🙏'];
-
-void _showMessageActions({
-  required BuildContext context,
-  required WidgetRef ref,
-  required TimelineMessage message,
-  required String channelId,
-  required bool isOwnMessage,
-  List<TimelineMessage>? allMessages,
-  String? currentPubkey,
-  bool isMember = false,
-  bool isArchived = false,
-}) {
-  showModalBottomSheet<void>(
-    context: context,
-    showDragHandle: true,
-    builder: (sheetContext) => SafeArea(
-      child: Padding(
-        padding: const EdgeInsets.fromLTRB(Grid.xs, 0, Grid.xs, Grid.xs),
-        child: Column(
-          mainAxisSize: MainAxisSize.min,
-          children: [
-            // Quick emoji row
-            Row(
-              mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-              children: [
-                for (final emoji in _quickEmojis)
-                  GestureDetector(
-                    onTap: () {
-                      Navigator.of(sheetContext).pop();
-                      ref
-                          .read(channelActionsProvider)
-                          .addReaction(message.id, emoji);
-                    },
-                    child: Container(
-                      width: 44,
-                      height: 44,
-                      alignment: Alignment.center,
-                      decoration: BoxDecoration(
-                        color: Theme.of(
-                          sheetContext,
-                        ).colorScheme.surfaceContainerHighest,
-                        shape: BoxShape.circle,
-                      ),
-                      child: Text(emoji, style: const TextStyle(fontSize: 20)),
-                    ),
-                  ),
-                GestureDetector(
-                  onTap: () {
-                    Navigator.of(sheetContext).pop();
-                    showEmojiPicker(
-                      context: context,
-                      onSelect: (emoji) {
-                        ref
-                            .read(channelActionsProvider)
-                            .addReaction(message.id, emoji);
-                      },
-                    );
-                  },
-                  child: Container(
-                    width: 44,
-                    height: 44,
-                    alignment: Alignment.center,
-                    decoration: BoxDecoration(
-                      color: Theme.of(
-                        sheetContext,
-                      ).colorScheme.surfaceContainerHighest,
-                      shape: BoxShape.circle,
-                    ),
-                    child: Icon(
-                      LucideIcons.plus,
-                      size: 20,
-                      color: Theme.of(
-                        sheetContext,
-                      ).colorScheme.onSurfaceVariant,
-                    ),
-                  ),
-                ),
-              ],
-            ),
-            const SizedBox(height: Grid.xs),
-            if (allMessages != null)
-              ListTile(
-                leading: const Icon(LucideIcons.messageSquareReply),
-                title: const Text('Reply in thread'),
-                onTap: () {
-                  Navigator.of(sheetContext).pop();
-                  Navigator.of(context).push(
-                    MaterialPageRoute<void>(
-                      builder: (_) => ThreadDetailPage(
-                        threadHead: message,
-                        allMessages: allMessages,
-                        channelId: channelId,
-                        currentPubkey: currentPubkey,
-                        isMember: isMember,
-                        isArchived: isArchived,
-                      ),
-                    ),
-                  );
-                },
-              ),
-            ListTile(
-              leading: const Icon(LucideIcons.copy),
-              title: const Text('Copy text'),
-              onTap: () {
-                Navigator.of(sheetContext).pop();
-                // Copy to clipboard
-                final data = ClipboardData(text: message.content);
-                Clipboard.setData(data);
-              },
-            ),
-            if (isOwnMessage) ...[
-              ListTile(
-                leading: const Icon(LucideIcons.pencil),
-                title: const Text('Edit message'),
-                onTap: () {
-                  Navigator.of(sheetContext).pop();
-                  _showEditSheet(
-                    context: context,
-                    ref: ref,
-                    message: message,
-                    channelId: channelId,
-                  );
-                },
-              ),
-              ListTile(
-                leading: Icon(
-                  LucideIcons.trash2,
-                  color: Theme.of(sheetContext).colorScheme.error,
-                ),
-                title: Text(
-                  'Delete message',
-                  style: TextStyle(
-                    color: Theme.of(sheetContext).colorScheme.error,
-                  ),
-                ),
-                onTap: () {
-                  Navigator.of(sheetContext).pop();
-                  _confirmDelete(
-                    context: context,
-                    ref: ref,
-                    messageId: message.id,
-                  );
-                },
-              ),
-            ],
-          ],
-        ),
-      ),
-    ),
-  );
-}
-
-void _showEditSheet({
-  required BuildContext context,
-  required WidgetRef ref,
-  required TimelineMessage message,
-  required String channelId,
-}) {
-  final controller = TextEditingController(text: message.content);
-  showModalBottomSheet<void>(
-    context: context,
-    isScrollControlled: true,
-    showDragHandle: true,
-    builder: (sheetContext) => Padding(
-      padding: EdgeInsets.fromLTRB(
-        Grid.xs,
-        0,
-        Grid.xs,
-        MediaQuery.viewInsetsOf(sheetContext).bottom + Grid.xs,
-      ),
-      child: SafeArea(
-        top: false,
-        child: Column(
-          mainAxisSize: MainAxisSize.min,
-          children: [
-            TextField(
-              controller: controller,
-              autofocus: true,
-              minLines: 1,
-              maxLines: 5,
-              decoration: const InputDecoration(hintText: 'Edit message'),
-            ),
-            const SizedBox(height: Grid.xxs),
-            Row(
-              mainAxisAlignment: MainAxisAlignment.end,
-              children: [
-                TextButton(
-                  onPressed: () => Navigator.of(sheetContext).pop(),
-                  child: const Text('Cancel'),
-                ),
-                const SizedBox(width: Grid.half),
-                FilledButton(
-                  onPressed: () {
-                    final text = controller.text.trim();
-                    if (text.isEmpty || text == message.content) {
-                      Navigator.of(sheetContext).pop();
-                      return;
-                    }
-                    ref
-                        .read(channelActionsProvider)
-                        .editMessage(
-                          channelId: channelId,
-                          eventId: message.id,
-                          content: text,
-                        );
-                    Navigator.of(sheetContext).pop();
-                  },
-                  child: const Text('Save'),
-                ),
-              ],
-            ),
-          ],
-        ),
-      ),
-    ),
-  );
-}
-
-void _confirmDelete({
-  required BuildContext context,
-  required WidgetRef ref,
-  required String messageId,
-}) {
-  showDialog<void>(
-    context: context,
-    builder: (dialogContext) => AlertDialog(
-      title: const Text('Delete message'),
-      content: const Text('This cannot be undone.'),
-      actions: [
-        TextButton(
-          onPressed: () => Navigator.of(dialogContext).pop(),
-          child: const Text('Cancel'),
-        ),
-        FilledButton(
-          onPressed: () {
-            Navigator.of(dialogContext).pop();
-            ref.read(channelActionsProvider).deleteMessage(messageId);
-          },
-          style: FilledButton.styleFrom(
-            backgroundColor: Theme.of(dialogContext).colorScheme.error,
-          ),
-          child: const Text('Delete'),
-        ),
-      ],
-    ),
-  );
-}
-
-// ---------------------------------------------------------------------------
 // Channel management
 // ---------------------------------------------------------------------------
 
@@ -1107,583 +855,6 @@ class _ReadOnlyNotice extends StatelessWidget {
           color: context.colors.onSurfaceVariant,
         ),
         textAlign: TextAlign.center,
-      ),
-    );
-  }
-}
-
-class _MembersSheet extends HookConsumerWidget {
-  final Channel channel;
-  final String? currentPubkey;
-
-  const _MembersSheet({required this.channel, required this.currentPubkey});
-
-  @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    final membersAsync = ref.watch(channelMembersProvider(channel.id));
-    final allMembers = membersAsync.asData?.value ?? const <ChannelMember>[];
-    final people = allMembers.where((member) => !member.isBot).toList();
-    final userCache = ref.watch(userCacheProvider);
-
-    // Determine if the current user can manage members.
-    final currentMember = allMembers.cast<ChannelMember?>().firstWhere(
-      (m) => m!.pubkey.toLowerCase() == currentPubkey?.toLowerCase(),
-      orElse: () => null,
-    );
-    final canManage =
-        currentMember != null &&
-        currentMember.isElevated &&
-        !channel.isArchived;
-
-    // Preload profiles for all members so avatars appear.
-    useEffect(() {
-      if (people.isNotEmpty) {
-        ref
-            .read(userCacheProvider.notifier)
-            .preload(people.map((m) => m.pubkey).toList());
-      }
-      return null;
-    }, [people.length]);
-
-    return Padding(
-      padding: EdgeInsets.fromLTRB(
-        Grid.xs,
-        0,
-        Grid.xs,
-        MediaQuery.viewInsetsOf(context).bottom + Grid.xs,
-      ),
-      child: SafeArea(
-        top: false,
-        child: SingleChildScrollView(
-          child: Column(
-            mainAxisSize: MainAxisSize.min,
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: [
-              Text('Members', style: context.textTheme.titleMedium),
-              const SizedBox(height: Grid.xxs),
-              Text(
-                'People in ${channel.displayLabel(currentPubkey: currentPubkey)}.',
-                style: context.textTheme.bodySmall?.copyWith(
-                  color: context.colors.onSurfaceVariant,
-                ),
-              ),
-              if (!channel.isDm) ...[const Divider(height: Grid.sm)],
-              SizedBox(
-                height: 280,
-                child: membersAsync.when(
-                  data: (_) => people.isEmpty
-                      ? Center(
-                          child: Text(
-                            'No people found.',
-                            style: context.textTheme.bodySmall?.copyWith(
-                              color: context.colors.outline,
-                            ),
-                          ),
-                        )
-                      : ListView(
-                          shrinkWrap: true,
-                          children: [
-                            for (final member in people)
-                              _MemberTile(
-                                member: member,
-                                currentPubkey: currentPubkey,
-                                profile: userCache[member.pubkey.toLowerCase()],
-                                canManage: canManage,
-                                isSelf:
-                                    member.pubkey.toLowerCase() ==
-                                    currentPubkey?.toLowerCase(),
-                                channelId: channel.id,
-                              ),
-                          ],
-                        ),
-                  loading: () =>
-                      const Center(child: CircularProgressIndicator()),
-                  error: (error, _) => Center(
-                    child: Text(
-                      error.toString(),
-                      style: context.textTheme.bodySmall?.copyWith(
-                        color: context.colors.error,
-                      ),
-                    ),
-                  ),
-                ),
-              ),
-            ],
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-const _changeableRoles = ['admin', 'member', 'guest'];
-
-class _MemberTile extends ConsumerWidget {
-  final ChannelMember member;
-  final String? currentPubkey;
-  final UserProfile? profile;
-  final bool canManage;
-  final bool isSelf;
-  final String channelId;
-
-  const _MemberTile({
-    required this.member,
-    required this.currentPubkey,
-    required this.profile,
-    required this.canManage,
-    required this.isSelf,
-    required this.channelId,
-  });
-
-  @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    final label = member.labelFor(currentPubkey);
-    final initial = label.substring(0, 1).toUpperCase();
-    final showMenu = canManage && !isSelf && !member.isOwner;
-
-    return ListTile(
-      contentPadding: EdgeInsets.zero,
-      leading: _MemberAvatar(avatarUrl: profile?.avatarUrl, initial: initial),
-      title: Text(label),
-      subtitle: Text(
-        _roleLabel(member.role),
-        style: context.textTheme.bodySmall?.copyWith(
-          color: context.colors.outline,
-        ),
-      ),
-      trailing: showMenu
-          ? IconButton(
-              icon: const Icon(LucideIcons.ellipsis, size: 18),
-              onPressed: () => _showMemberActions(context, ref),
-              visualDensity: VisualDensity.compact,
-            )
-          : null,
-    );
-  }
-
-  String _roleLabel(String role) {
-    if (role.isEmpty) return 'Member';
-    return '${role[0].toUpperCase()}${role.substring(1)}';
-  }
-
-  void _showMemberActions(BuildContext context, WidgetRef ref) {
-    final label = member.labelFor(currentPubkey);
-    showModalBottomSheet<void>(
-      context: context,
-      showDragHandle: true,
-      builder: (_) => SafeArea(
-        child: Column(
-          mainAxisSize: MainAxisSize.min,
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            Padding(
-              padding: const EdgeInsets.symmetric(horizontal: Grid.xs),
-              child: Text(label, style: context.textTheme.titleSmall),
-            ),
-            const SizedBox(height: Grid.xxs),
-            Padding(
-              padding: const EdgeInsets.symmetric(horizontal: Grid.xs),
-              child: Text(
-                'Change role',
-                style: context.textTheme.labelMedium?.copyWith(
-                  color: context.colors.outline,
-                ),
-              ),
-            ),
-            const SizedBox(height: Grid.half),
-            for (final role in _changeableRoles)
-              ListTile(
-                title: Text(_roleLabel(role)),
-                trailing: role == member.role
-                    ? Icon(
-                        LucideIcons.check,
-                        size: 16,
-                        color: context.colors.primary,
-                      )
-                    : null,
-                enabled: role != member.role,
-                onTap: role == member.role
-                    ? null
-                    : () async {
-                        Navigator.of(context).pop();
-                        await ref
-                            .read(channelActionsProvider)
-                            .changeMemberRole(
-                              channelId: channelId,
-                              pubkey: member.pubkey,
-                              role: role,
-                            );
-                      },
-              ),
-            const Divider(),
-            ListTile(
-              leading: Icon(
-                LucideIcons.userMinus,
-                size: 18,
-                color: context.colors.error,
-              ),
-              title: Text(
-                'Remove from channel',
-                style: TextStyle(color: context.colors.error),
-              ),
-              onTap: () async {
-                Navigator.of(context).pop();
-                final confirmed = await showDialog<bool>(
-                  context: context,
-                  builder: (context) => AlertDialog(
-                    title: const Text('Remove member'),
-                    content: Text('Remove $label from this channel?'),
-                    actions: [
-                      TextButton(
-                        onPressed: () => Navigator.of(context).pop(false),
-                        child: const Text('Cancel'),
-                      ),
-                      TextButton(
-                        onPressed: () => Navigator.of(context).pop(true),
-                        child: Text(
-                          'Remove',
-                          style: TextStyle(color: context.colors.error),
-                        ),
-                      ),
-                    ],
-                  ),
-                );
-                if (confirmed == true) {
-                  await ref
-                      .read(channelActionsProvider)
-                      .removeMember(
-                        channelId: channelId,
-                        pubkey: member.pubkey,
-                      );
-                }
-              },
-            ),
-            const SizedBox(height: Grid.xxs),
-          ],
-        ),
-      ),
-    );
-  }
-}
-
-class _MemberAvatar extends HookWidget {
-  final String? avatarUrl;
-  final String initial;
-
-  const _MemberAvatar({required this.avatarUrl, required this.initial});
-
-  @override
-  Widget build(BuildContext context) {
-    final failed = useState(false);
-
-    useEffect(() {
-      failed.value = false;
-      return null;
-    }, [avatarUrl]);
-
-    final url = avatarUrl;
-    if (url == null || failed.value) {
-      return CircleAvatar(child: Text(initial));
-    }
-    return CircleAvatar(
-      backgroundImage: NetworkImage(url),
-      onBackgroundImageError: (_, _) => failed.value = true,
-      child: null,
-    );
-  }
-}
-
-class _ManageChannelSheet extends HookConsumerWidget {
-  final Channel channel;
-
-  const _ManageChannelSheet({required this.channel});
-
-  @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    final canvasAsync = ref.watch(channelCanvasProvider(channel.id));
-    final isEditingCanvas = useState(false);
-    final isSavingCanvas = useState(false);
-    final isBusy = useState(false);
-    final actionError = useState<String?>(null);
-    final canvasController = useTextEditingController();
-
-    useEffect(() {
-      final canvas = canvasAsync.asData?.value;
-      if (!isEditingCanvas.value) {
-        canvasController.text = canvas?.content ?? '';
-      }
-      return null;
-    }, [canvasAsync.asData?.value.content, isEditingCanvas.value]);
-
-    final canJoin =
-        channel.visibility == 'open' &&
-        !channel.isArchived &&
-        !channel.isMember &&
-        !channel.isDm;
-    final canLeave = channel.isMember && !channel.isArchived && !channel.isDm;
-    final canEditCanvas = channel.isMember && !channel.isArchived;
-
-    Future<void> joinChannel() async {
-      if (isBusy.value) return;
-      isBusy.value = true;
-      actionError.value = null;
-      try {
-        await ref.read(channelActionsProvider).joinChannel(channel.id);
-        if (context.mounted) {
-          Navigator.of(context).pop(false);
-        }
-      } catch (error) {
-        actionError.value = error.toString();
-      } finally {
-        isBusy.value = false;
-      }
-    }
-
-    Future<void> leaveChannel() async {
-      if (isBusy.value) return;
-      isBusy.value = true;
-      actionError.value = null;
-      try {
-        await ref.read(channelActionsProvider).leaveChannel(channel.id);
-        if (context.mounted) {
-          Navigator.of(context).pop(true);
-        }
-      } catch (error) {
-        actionError.value = error.toString();
-      } finally {
-        isBusy.value = false;
-      }
-    }
-
-    Future<void> saveCanvas() async {
-      if (isSavingCanvas.value) {
-        return;
-      }
-      isSavingCanvas.value = true;
-      actionError.value = null;
-      try {
-        await ref
-            .read(channelActionsProvider)
-            .setCanvas(
-              channelId: channel.id,
-              content: canvasController.text.trim(),
-            );
-        if (context.mounted) {
-          isEditingCanvas.value = false;
-        }
-      } catch (error) {
-        actionError.value = error.toString();
-      } finally {
-        isSavingCanvas.value = false;
-      }
-    }
-
-    return Padding(
-      padding: EdgeInsets.fromLTRB(
-        Grid.xs,
-        0,
-        Grid.xs,
-        MediaQuery.viewInsetsOf(context).bottom + Grid.xs,
-      ),
-      child: SafeArea(
-        top: false,
-        child: ListView(
-          shrinkWrap: true,
-          children: [
-            Text('Manage channel', style: context.textTheme.titleMedium),
-            const SizedBox(height: Grid.xxs),
-            Text(
-              'Basic management for ${channel.name}.',
-              style: context.textTheme.bodySmall?.copyWith(
-                color: context.colors.onSurfaceVariant,
-              ),
-            ),
-            if (actionError.value case final error?) ...[
-              const SizedBox(height: Grid.xs),
-              Text(
-                error,
-                style: context.textTheme.bodySmall?.copyWith(
-                  color: context.colors.error,
-                ),
-              ),
-            ],
-            if (canJoin || canLeave) ...[
-              const SizedBox(height: Grid.xs),
-              Wrap(
-                spacing: Grid.xxs,
-                children: [
-                  if (canJoin)
-                    FilledButton.tonal(
-                      onPressed: isBusy.value ? null : joinChannel,
-                      child: Text(isBusy.value ? 'Joining…' : 'Join channel'),
-                    ),
-                  if (canLeave)
-                    OutlinedButton(
-                      onPressed: isBusy.value ? null : leaveChannel,
-                      child: Text(isBusy.value ? 'Leaving…' : 'Leave channel'),
-                    ),
-                ],
-              ),
-            ],
-            const SizedBox(height: Grid.sm),
-            Text('Context', style: context.textTheme.labelLarge),
-            const SizedBox(height: Grid.xxs),
-            _ContextCard(
-              label: 'Description',
-              value: channel.description,
-              emptyLabel: 'No description set',
-            ),
-            const SizedBox(height: Grid.xxs),
-            _ContextCard(
-              label: 'Topic',
-              value: channel.topic,
-              emptyLabel: 'No topic set',
-            ),
-            const SizedBox(height: Grid.xxs),
-            _ContextCard(
-              label: 'Purpose',
-              value: channel.purpose,
-              emptyLabel: 'No purpose set',
-            ),
-            if (!channel.isDm) ...[
-              const SizedBox(height: Grid.sm),
-              Text('Canvas', style: context.textTheme.labelLarge),
-              const SizedBox(height: Grid.xxs),
-              canvasAsync.when(
-                data: (canvas) {
-                  if (isEditingCanvas.value) {
-                    return Column(
-                      crossAxisAlignment: CrossAxisAlignment.start,
-                      children: [
-                        TextField(
-                          controller: canvasController,
-                          maxLines: 8,
-                          minLines: 6,
-                          decoration: const InputDecoration(
-                            hintText: 'Write your canvas content in Markdown…',
-                          ),
-                        ),
-                        const SizedBox(height: Grid.xxs),
-                        Row(
-                          mainAxisAlignment: MainAxisAlignment.end,
-                          children: [
-                            TextButton(
-                              onPressed: isSavingCanvas.value
-                                  ? null
-                                  : () {
-                                      isEditingCanvas.value = false;
-                                      canvasController.text =
-                                          canvas.content ?? '';
-                                    },
-                              child: const Text('Cancel'),
-                            ),
-                            const SizedBox(width: Grid.half),
-                            FilledButton(
-                              onPressed: isSavingCanvas.value
-                                  ? null
-                                  : saveCanvas,
-                              child: Text(
-                                isSavingCanvas.value
-                                    ? 'Saving…'
-                                    : 'Save canvas',
-                              ),
-                            ),
-                          ],
-                        ),
-                      ],
-                    );
-                  }
-
-                  return Column(
-                    crossAxisAlignment: CrossAxisAlignment.start,
-                    children: [
-                      Container(
-                        width: double.infinity,
-                        padding: const EdgeInsets.all(Grid.xs),
-                        decoration: BoxDecoration(
-                          color: context.colors.surfaceContainerHighest,
-                          borderRadius: BorderRadius.circular(Radii.md),
-                        ),
-                        child: Text(
-                          canvas.content?.trim().isNotEmpty == true
-                              ? canvas.content!
-                              : 'No canvas set for this channel.',
-                          style: context.textTheme.bodyMedium?.copyWith(
-                            color: context.colors.onSurfaceVariant,
-                          ),
-                        ),
-                      ),
-                      const SizedBox(height: Grid.xxs),
-                      Align(
-                        alignment: Alignment.centerRight,
-                        child: FilledButton.tonal(
-                          onPressed: canEditCanvas
-                              ? () => isEditingCanvas.value = true
-                              : null,
-                          child: Text(
-                            canvas.content?.trim().isNotEmpty == true
-                                ? 'Edit canvas'
-                                : 'Create canvas',
-                          ),
-                        ),
-                      ),
-                    ],
-                  );
-                },
-                loading: () => const Center(child: CircularProgressIndicator()),
-                error: (error, _) => Text(
-                  error.toString(),
-                  style: context.textTheme.bodySmall?.copyWith(
-                    color: context.colors.error,
-                  ),
-                ),
-              ),
-            ],
-          ],
-        ),
-      ),
-    );
-  }
-}
-
-class _ContextCard extends StatelessWidget {
-  final String label;
-  final String? value;
-  final String emptyLabel;
-
-  const _ContextCard({
-    required this.label,
-    required this.value,
-    required this.emptyLabel,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      width: double.infinity,
-      padding: const EdgeInsets.all(Grid.xs),
-      decoration: BoxDecoration(
-        color: context.colors.surfaceContainerHighest,
-        borderRadius: BorderRadius.circular(Radii.md),
-      ),
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          Text(
-            label,
-            style: context.textTheme.labelSmall?.copyWith(
-              color: context.colors.outline,
-              fontWeight: FontWeight.w600,
-            ),
-          ),
-          const SizedBox(height: Grid.half),
-          Text(
-            value?.trim().isNotEmpty == true ? value!.trim() : emptyLabel,
-            style: context.textTheme.bodyMedium?.copyWith(
-              color: context.colors.onSurfaceVariant,
-            ),
-          ),
-        ],
       ),
     );
   }
